Add principal modules and dependency inversion injected vars
Start a project with common modules used in aws infraestructure such as
sns, iam, s3 and lambda and add dependency inversion to have cohesion
between components without hardcoded strings.

Is highttly recommended not add default values for required vars,
instead of it is better to enforces to inject that vars.